SELENA/Python/Scripts/requirements.txt
pvporcupine
pvrecorder
zmq
pydub
pvrhino
google-cloud-texttospeech
simpleaudio
lcu-driver
